// Frontend/src/app/models/email-details.ts
export class EmailDetails {
    EmailToID: string = '';
    EmailCC: string = '';
    EmailSubject: string = '';
    Application: string = '';
    Env: string = '';
    Date: string = '';
    Hour: string = '';
    Duration: string = '';
    Downtime: string = '';
    Contents: string = '';
    Tests: string = '';
    AddOn: string = '';
    Title: string = '';
    IsEmerg: string = '';
    Status: string = '';
    Type: string = '';
}